Refactor configuration
We merge version, program name and the rest of the config with the
existing config. Instead of using the template for the version, we now
use resource filtering and an Apache composite configuration:

The existing configuration from the user's home directory and one from
classpath that is resource filtered and contains version and program
name. The latter is the actual project.name from the POM.